+++ example/invaders/model/pieces.dart
@@ -5,8 +5,6 @@ class Spaceship extends MovablePiece {
     randomInit();
     width = 70;
     height = 70;
-    isCovered = false;
-    isTagged = false;
     shape = PieceShape.IMG;
     imgId = 'spaceship';
     videoId = 'invader';
@@ -19,8 +17,6 @@ class Laser extends MovablePiece {
     randomInit();
     width = 4;
     height = 50;
-    isCovered = false;
-    isTagged = false;
     color.main = 'gray';
     color.border = 'red';
     speed.dy = 6;
@@ -36,8 +32,6 @@ class Cloud extends MovablePiece {
     randomInit();
     width = 80;
     height = 56;
-    isCovered = false;
-    isTagged = false;
     shape = PieceShape.IMG;
     imgId = 'cloud';
   }
@@ -45,7 +39,7 @@ class Cloud extends MovablePiece {
 
 class Creature extends MovablePiece {  
   Creature(int id) {
-    randomInit();
+    randomExtraInit();
     width = 48;
     height = 64;
     if (dy < 2) {